requirements.txt
constructs>=10.0.0,<11.0.0
Flask~=2.0.3
PyYAML~=6.0.2